+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - add `MOODS_NAME_ONLY` output to `VrchAudioMusic2EmotionNode`
+- add `TEXT Word Replacer @ vrch.ai` for rule-based text replacement
 
 ### Fixed